Add files via upload
It asks for the ip address subnet in CIDR format and the name of the file to be saved. It displays a basic percentage counter on screen to track how much of the subnet has been scanned. It saves only the online hosts on the file specified. The script can be stopped by pressing Ctrl+c. For now, it's just a basic subnet pinger.